+++ Runtime/Scripts/Tezos/Managers/TezosAuthenticator.cs
@@ -11,7 +11,6 @@ namespace TezosSDK.Tezos
 		[SerializeField] private GameObject deepLinkButton;
 		[SerializeField] private GameObject socialLoginButton;
 		[SerializeField] private GameObject logoutButton;
-		[SerializeField] private Image darkBG;
 
 		// Platform flags to determine the current running platform
 		private bool _isMobile;
@@ -41,18 +40,18 @@ namespace TezosSDK.Tezos
 			}
 		}
 
-		private void Update()
-		{
-			if (Input.GetKeyDown(KeyCode.Return))
-			{
-				Tezos.Wallet.Connect(WalletProviderType.beacon);
-			}
-
-			if (Input.GetKeyDown(KeyCode.Escape))
-			{
-				Tezos.Wallet.Disconnect();
-			}
-		}
+		// private void Update()
+		// {
+		// 	if (Input.GetKeyDown(KeyCode.Return))
+		// 	{
+		// 		Tezos.Wallet.Connect(WalletProviderType.beacon);
+		// 	}
+		//
+		// 	if (Input.GetKeyDown(KeyCode.Escape))
+		// 	{
+		// 		Tezos.Wallet.Disconnect();
+		// 	}
+		// }
 
 		private void OnEnable()
 		{
@@ -83,6 +82,12 @@ namespace TezosSDK.Tezos
 			ToggleUIElements(true);
 		}
 
+		private void OnWalletDisconnected(WalletInfo walletInfo)
+		{
+			Logger.LogDebug("TezosAuthenticator.OnWalletDisconnected");
+			ToggleUIElements(false);
+		}
+
 		public void ConnectByDeeplink()
 		{
 			Tezos.Wallet.Connect(WalletProviderType.beacon);
@@ -111,6 +116,7 @@ namespace TezosSDK.Tezos
 			// Subscribe to wallet events for handling user authentication.
 			Tezos.Wallet.EventManager.HandshakeReceived += OnHandshakeReceived;
 			Tezos.Wallet.EventManager.WalletConnected += OnWalletConnected;
+			Tezos.Wallet.EventManager.WalletDisconnected += OnWalletDisconnected;
 		}
 
 		/// <summary>
@@ -124,7 +130,6 @@ namespace TezosSDK.Tezos
 				deepLinkButton.SetActive(false);
 				socialLoginButton.SetActive(false);
 				qrCodeGenerator.gameObject.SetActive(false);
-				darkBG.gameObject.SetActive(false);
 			}
 			else
 			{
@@ -136,8 +141,6 @@ namespace TezosSDK.Tezos
 
 				// Activate qrCodePanel only on standalone and not authenticated
 				qrCodeGenerator.gameObject.SetActive(!_isMobile && !_isWebGL);
-
-				darkBG.gameObject.SetActive(true);
 			}
 
 			logoutButton.SetActive(isAuthenticated);
@@ -147,6 +150,7 @@ namespace TezosSDK.Tezos
 		{
 			TezosManager.Instance.Wallet.EventManager.HandshakeReceived -= OnHandshakeReceived;
 			TezosManager.Instance.Wallet.EventManager.WalletConnected -= OnWalletConnected;
+			TezosManager.Instance.Wallet.EventManager.WalletDisconnected -= OnWalletDisconnected;
 		}
 	}